requirements.txt
datasets==4.5.0
duckdb==1.5.0
faiss-cpu==1.13.2
lxml==6.0.2
networkx==3.6.1
numpy==2.4.3
pandas==3.0.1
pypika==0.51.1
python_dateutil==2.9.0.post0
rapidfuzz==3.14.3
Requests==2.32.5
scikit_learn==1.8.0
sentence_transformers==5.2.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
